requirements.txt
requests >= 2
pandas == 2.2.3
entsoe-py == 0.6.16